Update on "qrelu6 benchmarking"
Differential Revision: [D18319499](https://our.internmc.facebook.com/intern/diff/D18319499)

+++ benchmarks/operator_benchmark/pt/qactivation_test.py
@@ -25,7 +25,7 @@ qactivation_configs = op_bench.cross_product_configs(
     permute_dims=(False, True),
     inplace=(False, True),
     dtype=(torch.quint8, torch.qint8, torch.qint32),
-    tags=('short',)
+    tags=('long',)
 )
